Tidying up the code
Created subfolder structure and moved some of the functionality there.
Changed the formating of JSX returns in react-components, so the browser stops complaining about react-keys.
Changed InfoBox.js to be a bit more easier to change by adding formatBox function that can be used to change the styling of all boxes.
Wrote some comments to functions that don't seem so self eveident.
Lots of little fixes and cleanups around the code.